Renaming jest fixtures
