+++ .jest/__mocks__/@utils/accessDetailsAndPricing.ts
@@ -1,3 +1,3 @@
-import { assets } from '../../__fixtures__/assetsWithAccessDetails'
+import { assets } from '../../__fixtures__/datasetsWithAccessDetails'
 
 export const getAccessDetailsForAssets = jest.fn().mockResolvedValue(assets)
