+++ src/components/@shared/AssetTeaser/index.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import testRender from '../../../../.jest/testRender'
 import AssetTeaser from './index'
-import { asset } from '../../../../.jest/__fixtures__/assetWithAccessDetails'
+import { asset } from '../../../../.jest/__fixtures__/datasetWithAccessDetails'
 
 describe('@shared/AssetTeaser', () => {
   testRender(<AssetTeaser asset={asset} />)
