+++ src/components/@shared/Price/index.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react'
 import React from 'react'
 import Price from './index'
-import { asset } from '../../../../.jest/__fixtures__/assetWithAccessDetails'
+import { asset } from '../../../../.jest/__fixtures__/datasetWithAccessDetails'
 import prices from '../../../../.jest/__fixtures__/prices'
 
 jest.mock('../../../@context/Prices', () => ({
