+++ src/components/Asset/AssetContent/Bookmark.test.tsx
@@ -1,11 +1,11 @@
 import { render, screen } from '@testing-library/react'
 import React from 'react'
 import Bookmark from './Bookmark'
-import { assetAquarius } from '../../../../.jest/__fixtures__/assetAquarius'
+import { datasetAquarius } from '../../../../.jest/__fixtures__/datasetAquarius'
 
 describe('src/components/Asset/AssetContent/Bookmark.tsx', () => {
   it('renders Metadata for a data Asset', () => {
-    render(<Bookmark did={assetAquarius.id} />)
+    render(<Bookmark did={datasetAquarius.id} />)
     expect(screen.getByTitle('Add Bookmark')).toBeInTheDocument()
   })
 })
