+++ src/components/Asset/AssetContent/MetaFull.test.tsx
@@ -1,12 +1,12 @@
 import { render, screen } from '@testing-library/react'
 import React from 'react'
 import MetaFull from './MetaFull'
-import { assetAquarius } from '../../../../.jest/__fixtures__/assetAquarius'
+import { datasetAquarius } from '../../../../.jest/__fixtures__/datasetAquarius'
 import { algorithmAquarius } from '../../../../.jest/__fixtures__/algorithmAquarius'
 
 describe('src/components/Asset/AssetContent/MetaFull.tsx', () => {
   it('renders Metadata for a data Asset', () => {
-    render(<MetaFull ddo={assetAquarius} />)
+    render(<MetaFull ddo={datasetAquarius} />)
     expect(screen.getByText('DID')).toBeInTheDocument()
   })
   it('renders Metadata for an algorithm', () => {
